+++ grunt/browser_clients.js
@@ -29,9 +29,9 @@ module.exports = function (grunt) {
     '_check_for_confirmation',
     '_upload_archive:release',
     'browser_clients:build',
-    'run:init_submodules',
+    'run:clone_bower_repo',
     'copy:dist_to_bower',
-    'run:release_bower_subm_tag'
+    'run:release_bower_tag'
   ]);
 
   grunt.registerTask('_upload_archive', function (type) {