+++ grunt/config/run.js
@@ -10,7 +10,9 @@ var esOpts = [
   '-D es.logger.level=ERROR'
 ].join(' ');
 
-module.exports = {
+var utils = require('../utils');
+
+var config = {
   generate: {
     exec: 'node ./scripts/generate/index.js',
     options: {
@@ -27,39 +29,36 @@ module.exports = {
       ready: /listening/
     }
   },
-  install_es_master: {
-    exec: './scripts/es.sh install master',
-  },
-  es_master: {
-    exec: './.snapshots/master_nightly/bin/elasticsearch ' + esOpts,
-    options: {
-      wait: false,
-      quiet: true,
-      onClose: function () {
-
-      },
-      onReady: function () {
-
-      }
-    }
-  },
-  'install_es_0.90': {
-    exec: './scripts/es.sh install 0.90',
-  },
-  'es_0.90': {
-    exec: './.snapshots/0.90_nightly/bin/elasticsearch -f ' + esOpts,
-    options: {
-      wait: false,
-      quiet: true
-    }
-  },
-  init_submodules: {
-    exec: 'git submodule update --init',
+  clone_bower_repo: {
+    exec: [
+      'test -d src/bower_es_js',
+      'git clone git@github.com:elasticsearch/bower-elasticsearch-js.git <%= bowerSubmodule %>'
+    ].join(' || '),
     options: {
       quiet: true
     }
   },
-  release_bower_subm_tag: {
+  release_bower_tag: {
     exec: 'node ./scripts/release/bower'
   }
 };
+
+utils.branches.forEach(function (branch) {
+
+  config['install_es_' + branch] = {
+    exec: './scripts/es.sh install ' + branch,
+  };
+
+  config['es_' + branch] = {
+    exec:
+      './.snapshots/' + branch + '_nightly/bin/elasticsearch ' +
+      (branch === '0.90' ? '-f ' : '') +
+      esOpts,
+    options: {
+      wait: false,
+      quiet: true
+    }
+  };
+});
+
+module.exports = config;