+++ scripts/release/bower.js
@@ -1,10 +1,10 @@
 var fs = require('fs');
 var spawn = require('../_spawn');
 var async = require('async');
-var _ = require('lodash');
+var _ = require('lodash-node');
 
 var root = require('path').join(__dirname, '../..');
-var bowerDir = root + '/src/bower-elasticsearch-js';
+var bowerDir = root + '/src/bower_es_js';
 
 // get both the bower and node package files
 var bowerJson = require(bowerDir + '/bower.json');