+++ src/lib/apis/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  'master': require('./master'),
+  '1.x': require('./1_x'),
+  '1.0': require('./1_0'),
+  '0.90': require('./0_90'),
+  _default: '0.90'
+};