+++ src/lib/client_action.js
@@ -5,7 +5,6 @@
 module.exports = ClientAction;
 
 var _ = require('./utils');
-var when = require('when');
 
 function ClientAction(spec) {
   if (!_.isPlainObject(spec.params)) {
@@ -26,12 +25,14 @@ function ClientAction(spec) {
     }
 
     try {
-      return exec(this.transport, spec, params, cb);
+      return exec(this.transport, spec, _.clone(params), cb);
     } catch (e) {
       if (typeof cb === 'function') {
         _.nextTick(cb, e);
       } else {
-        return when.reject(e);
+        var def = this.transport.defer();
+        def.reject(e);
+        return def.promise;
       }
     }
   }