+++ src/lib/transport.js
@@ -30,6 +30,9 @@ function Transport(config) {
   // setup max retries
   this.maxRetries = config.hasOwnProperty('maxRetries') ? config.maxRetries : 3;
 
+  // setup endpoint to use for sniffing
+  this.sniffEndpoint = config.hasOwnProperty('sniffEndpoint') ? config.sniffEndpoint : '/_nodes/_all/clear';
+
   // setup requestTimeout default
   this.requestTimeout = config.hasOwnProperty('requestTimeout') ? config.requestTimeout : 30000;
 
@@ -71,9 +74,9 @@ function Transport(config) {
   }
 
   if (config.sniffInterval) {
-    this._sniffTimeout = setTimeout(function doSniff() {
+    this._timeout(function doSniff() {
       self.sniff();
-      self._sniffTimeout = setTimeout(doSniff, config.sniffInterval);
+      self._timeout(doSniff, config.sniffInterval);
     }, config.sniffInterval);
   }
 
@@ -196,7 +199,7 @@ Transport.prototype.request = function (params, cb) {
       return;
     }
 
-    clearTimeout(requestTimeoutId);
+    self._timeout(requestTimeoutId);
     var parsedBody;
     var isJson = !headers || (headers['content-type'] && ~headers['content-type'].indexOf('application/json'));
 
@@ -258,14 +261,14 @@ Transport.prototype.request = function (params, cb) {
 
     aborted = true;
     remainingRetries = 0;
-    clearTimeout(requestTimeoutId);
+    self._timeout(requestTimeoutId);
     if (typeof requestAborter === 'function') {
       requestAborter();
     }
   }
 
   if (requestTimeout && requestTimeout !== Infinity) {
-    requestTimeoutId = setTimeout(function () {
+    requestTimeoutId = this._timeout(function () {
       respond(new errors.RequestTimeout('Request Timeout after ' + requestTimeout + 'ms'));
       abortRequest();
     }, requestTimeout);
@@ -292,7 +295,31 @@ Transport.prototype.request = function (params, cb) {
   return ret;
 };
 
+Transport.prototype._timeout = function (cb, delay) {
+  this._timers = this._timers || [];
+  var id;
 
+  if ('function' !== typeof cb) {
+    id = cb;
+    cb = void 0;
+  }
+
+  if (cb) {
+    // set the timer
+    id = setTimeout(cb, delay);
+    this._timers.push(id);
+    return id;
+  }
+
+  if (id) {
+    clearTimeout(id);
+
+    var i = this._timers.indexOf(id);
+    if (i !== -1) {
+      this._timers.splice(i, 1);
+    }
+  }
+};
 
 /**
  * Ask an ES node for a list of all the nodes, add/remove nodes from the connection
@@ -308,7 +335,7 @@ Transport.prototype.sniff = function (cb) {
   cb = typeof cb === 'function' ? cb : _.noop;
 
   this.request({
-    path: '/_cluster/nodes',
+    path: this.sniffEndpoint,
     method: 'GET'
   }, function (err, resp, status) {
     if (!err && resp && resp.nodes) {
@@ -327,5 +354,6 @@ Transport.prototype.sniff = function (cb) {
  */
 Transport.prototype.close = function () {
   this.log.close();
+  _.each(this._timers, clearTimeout);
   this.connectionPool.close();
 };