+++ test/mocks/browser_http.js
@@ -29,7 +29,7 @@
 
 module.exports = MockHttpRequest;
 
-var _ = require('lodash');
+var _ = require('lodash-node');
 
 function MockHttpRequest() {
   // These are internal flags and data structures