+++ test/unit/specs/client.js
@@ -1,8 +1,9 @@
 describe('Client instances creation', function () {
+  var path = require('path');
   var es = require('../../../src/elasticsearch');
-  var api = require('../../../src/lib/api');
-  var api090 = require('../../../src/lib/api_0_90');
+  var apis = require('../../../src/lib/apis');
   var expect = require('expect.js');
+  var stub = require('../../utils/auto_release_stub').make();
   var client;
 
   beforeEach(function () {
@@ -21,17 +22,17 @@ describe('Client instances creation', function () {
   });
 
   it('inherits the 0.90 API by default', function () {
-    expect(client.bulk).to.eql(api090.bulk);
-    expect(client.cluster.nodeStats).to.eql(api090.cluster.prototype.nodeStats);
+    expect(client.bulk).to.eql(apis['0.90'].bulk);
+    expect(client.cluster.nodeStats).to.eql(apis['0.90'].cluster.prototype.nodeStats);
   });
 
-  it('inherits the master API when specified', function () {
+  it('inherits the 1.0 API when specified', function () {
     client.close();
     client = es.Client({
-      apiVersion: 'master'
+      apiVersion: '1.0'
     });
-    expect(client.bulk).to.eql(api.bulk);
-    expect(client.nodes.stats).to.eql(api.nodes.prototype.stats);
+    expect(client.bulk).to.eql(apis['1.0'].bulk);
+    expect(client.nodes.stats).to.eql(apis['1.0'].nodes.prototype.stats);
   });
 
   it('closing the client causes it\'s transport to be closed', function () {
@@ -49,5 +50,14 @@ describe('Client instances creation', function () {
     expect(client.transport.log.listenerCount('info')).to.eql(0);
     expect(client.transport.log.listenerCount('debug')).to.eql(0);
     expect(client.transport.log.listenerCount('trace')).to.eql(0);
+  });
+
+  describe('#ping', function () {
+    it('sets the default requestTimeout to 100', function () {
+      stub(client.transport, 'request');
+      client.ping();
+      expect(client.transport.request.callCount).to.be(1);
+      expect(client.transport.request.lastCall.args[0].requestTimeout).to.be(100);
+    });
   });
 });