+++ test/unit/specs/round_robin_selector.js
@@ -1,6 +1,6 @@
 describe('Round Robin Selector', function () {
   var selector = require('../../../src/lib/selectors/round_robin');
-  var _ = require('lodash');
+  var _ = require('lodash-node');
   var expect = require('expect.js');
 
   it('chooses options in order', function () {